+++ src/components/burger-ingredients/hooks/index.ts
@@ -1,2 +1,1 @@
 export { useBurgerIngredientsTabs } from './use-burger-ingredients-tabs.ts';
-export { useBurgerIngredientsModal } from './use-burger-ingredients-modal.ts';
