--- src/components/burger-ingredients/hooks/use-burger-ingredients-modal.ts
@@ -1,13 +1,0 @@
-import { TIngredientItem } from '@/api';
-import { useDispatch } from 'react-redux';
-import { setSelectedIngredient, TAppDispatch } from '@/store';
-
-export const useBurgerIngredientsModal = () => {
-  const dispatch = useDispatch<TAppDispatch>();
-
-  const handleItemClick = (item: TIngredientItem) => {
-    dispatch(setSelectedIngredient(item));
-  };
-
-  return { handleItemClick };
-};
