+++ src/components/burger-ingredients/ui/burger-ingredients-details/burger-ingredients-details.tsx
@@ -3,45 +3,31 @@ import { Text } from '@/components';
 import styles from './burger-ingredients-details.module.scss';
 import classNames from 'classnames';
 import { BurgerIngredientsDetailsCharacteristics } from './ui';
-import { useSelector } from 'react-redux';
-import { TRootState } from '@/store';
-import { useParams } from 'react-router-dom';
+import { TIngredientItem } from '@/api';
 
 type TBurgerIngredientsDetailsProps = {
   asPage?: boolean;
+  item?: TIngredientItem;
 };
 
 export const BurgerIngredientsDetails: FC<TBurgerIngredientsDetailsProps> = ({
   asPage,
+  item,
 }) => {
-  const params = useParams();
-
-  const ingredients = useSelector(
-    (state: TRootState) => state.ingredients.ingredients,
-  );
-
-  const currentIngredient = ingredients.find(
-    (item) => item._id === params.ingredientId,
-  );
-
   return (
     <>
-      {currentIngredient && (
+      {item && (
         <div
           className={classNames(styles.content, asPage && styles.contentAsPage)}
         >
-          <img
-            className="pb-4"
-            src={currentIngredient.image_large}
-            alt="Ingredient image"
-          />
+          <img className="pb-4" src={item.image_large} alt="Ingredient image" />
           <Text
             className={classNames(styles.itemName, 'pb-8')}
             textType="medium"
           >
-            {currentIngredient.name}
+            {item.name}
           </Text>
-          <BurgerIngredientsDetailsCharacteristics item={currentIngredient} />
+          <BurgerIngredientsDetailsCharacteristics item={item} />
         </div>
       )}
     </>
