+++ src/components/burger-ingredients/ui/burger-ingredients-modal/burger-ingredients-modal.tsx
@@ -3,18 +3,25 @@ import { Modal } from '@/components';
 import classNames from 'classnames';
 import styles from './burger-ingredients-modal.module.scss';
 import { BurgerIngredientsDetails } from '../burger-ingredients-details';
-import { useDispatch } from 'react-redux';
-import { setSelectedIngredient, TAppDispatch } from '@/store';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
+import { useSelector } from 'react-redux';
+import { TRootState } from '@/store';
 
 export const BurgerIngredientsModal: FC = () => {
-  const dispatch = useDispatch<TAppDispatch>();
+  const params = useParams();
   const navigate = useNavigate();
 
   const handleModalClose = () => {
-    dispatch(setSelectedIngredient(null));
     navigate(-1);
   };
+
+  const ingredients = useSelector(
+    (state: TRootState) => state.ingredients.ingredients,
+  );
+
+  const currentIngredient = ingredients.find(
+    (item) => item._id === params.ingredientId,
+  );
 
   return (
     <Modal
@@ -22,7 +29,7 @@ export const BurgerIngredientsModal: FC = () => {
       title="Детали ингредиента"
       onClose={handleModalClose}
     >
-      <BurgerIngredientsDetails />
+      <BurgerIngredientsDetails item={currentIngredient} />
     </Modal>
   );
 };
