+++ src/components/home/hooks/use-ingredients.ts
@@ -1,9 +1,7 @@
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchIngredientsAsync, TAppDispatch, TRootState } from '@/store';
+import { useSelector } from 'react-redux';
+import { TRootState } from '@/store';
 
 export const useIngredients = () => {
-  const dispatch = useDispatch<TAppDispatch>();
   const isLoading = useSelector(
     (state: TRootState) => state.ingredients.isLoading,
   );
@@ -11,10 +9,6 @@ export const useIngredients = () => {
     (state: TRootState) => state.ingredients.errorMessage,
   );
   const isError = !!errorMessage;
-
-  useEffect(() => {
-    dispatch(fetchIngredientsAsync());
-  }, [dispatch]);
 
   return { isLoading, isError, errorMessage };
 };
