+++ src/store/actions/profile.action.ts
@@ -31,22 +31,23 @@ export const loginAsync =
 
 export const checkProfileAuthAsync =
   (): ThunkAction<void, TRootState, unknown, any> => async (dispatch) => {
-    if (accessTokenLocalStorage.get()) {
-      try {
-        dispatch(setProfileIsLoading(true));
-
-        const profile = await getProfile();
-
-        dispatch(setProfile(profile));
-      } catch (error) {
-        accessTokenLocalStorage.remove();
-        refreshTokenLocalStorage.remove();
-      } finally {
-        dispatch(setAuthChecked(true));
-        dispatch(setProfileIsLoading(false));
-      }
-    } else {
+    if (!accessTokenLocalStorage.get()) {
       dispatch(setAuthChecked(true));
+      return;
+    }
+
+    try {
+      dispatch(setProfileIsLoading(true));
+
+      const profile = await getProfile();
+
+      dispatch(setProfile(profile));
+    } catch (error) {
+      accessTokenLocalStorage.remove();
+      refreshTokenLocalStorage.remove();
+    } finally {
+      dispatch(setAuthChecked(true));
+      dispatch(setProfileIsLoading(false));
     }
   };
