+++ src/store/reducers/index.ts
@@ -1,9 +1,6 @@
 export { default as ingredientsReducer } from './ingredients.reducer.ts';
 export * from './ingredients.reducer.ts';
 
-export { default as selectedIngredientReducer } from './selected-ingredient.reducer.ts';
-export * from './selected-ingredient.reducer.ts';
-
 export { default as burgerConstructorReducer } from './burger-constructor.reducer.ts';
 export * from './burger-constructor.reducer.ts';
