--- src/store/reducers/selected-ingredient.reducer.ts
@@ -1,27 +1,0 @@
-import { TIngredientItem } from '@/api';
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-
-type TSelectedIngredientState = {
-  selectedIngredient: TIngredientItem | null;
-};
-
-const initialState: TSelectedIngredientState = {
-  selectedIngredient: null,
-};
-
-const selectedIngredientSlice = createSlice({
-  name: 'selectedIngredient',
-  initialState,
-  reducers: {
-    setSelectedIngredient(
-      state,
-      action: PayloadAction<TIngredientItem | null>,
-    ) {
-      state.selectedIngredient = action.payload;
-    },
-  },
-});
-
-export const { setSelectedIngredient } = selectedIngredientSlice.actions;
-
-export default selectedIngredientSlice.reducer;
