+++ src/store/store.ts
@@ -1,7 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import {
   ingredientsReducer,
-  selectedIngredientReducer,
   burgerConstructorReducer,
   currentOrderReducer,
   profileReducer,
@@ -10,7 +9,6 @@ import {
 export const store = configureStore({
   reducer: {
     ingredients: ingredientsReducer,
-    selectedIngredient: selectedIngredientReducer,
     burgerConstructor: burgerConstructorReducer,
     currentOrder: currentOrderReducer,
     profile: profileReducer,
